pyproject.toml
[project]
name = "deep-ai-research"
version = "0.1.0"
description = "Personal deep-research tool for AI-related topics. Codename: dair. Corpus + authority graph + Claude Code orchestrator."
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "feedparser>=6.0.11",      # RSS / Atom parsing
    "httpx>=0.27.0",           # HTTP client
    "pydantic>=2.7.0",         # frontmatter schema validation
    "pyyaml>=6.0.1",           # config files
    "python-frontmatter>=1.1.0", # markdown + YAML frontmatter handling
    "beautifulsoup4>=4.12.3",  # HTML cleanup
    "python-dateutil>=2.9.0",  # robust date parsing
    "click>=8.1.7",            # CLI
    "pysqlite3-binary>=0.5.4", # ships its own modern sqlite, bypasses old system sqlite
    "sqlite-vec>=0.1.6",       # vector extension
    "numpy>=1.26.0",           # embeddings + numpy fallback
    "praw>=7.7.1",             # Reddit ingestion (Tier 2)
]

[project.optional-dependencies]
embed = [
    # Local embedding model. Install with: uv sync --extra embed
    "sentence-transformers>=3.0.1",
    "onnxruntime>=1.18.0",
]
podcasts = [
    # Whisper transcription. Install with: uv sync --extra podcasts
    "faster-whisper>=1.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingest", "mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # ruff format handles line length

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
